finishing TODOs
