+++ src/app/(App)/bulkcore1/TimeSection.tsx
@@ -1,107 +1,23 @@
 import Border from '@/components/border/Border'
 import TimelineComponent from '@/components/timelineComp/TimelineComp'
 import BuyWalletStatus from '@/components/walletStatus/BuyWalletStatus'
-import { saleStatus } from '@/utils/broker'
-import { ApiPromise } from '@polkadot/api'
 import {
-  BrokerConstantsType,
+  useBrokerConstants,
+  useCurrentBlockNumber,
+  useSubstrateQuery,
+} from '@/hooks/useSubstrateQuery'
+import { saleStatus } from '@/utils/broker'
+import {
   ConfigurationType,
   SaleInfoType,
   StatusType,
   blockTimeToUTC,
-  getConstants,
   getCurrentBlockNumber,
   useBalance,
   useInkathon,
 } from '@poppyseed/lastic-sdk'
 import { useEffect, useMemo, useState } from 'react'
 import AnalyticSection from './AnalyticSection'
-
-// Define a type for the queryParams
-type QueryParams = (string | number | Record<string, unknown>)[]
-
-// Custom hook for querying substrate state
-function useSubstrateQuery(
-  api: ApiPromise | undefined,
-  queryKey: string,
-  queryParams: QueryParams = [],
-) {
-  const [data, setData] = useState<string | null>(null)
-
-  useEffect(() => {
-    const fetchData = async () => {
-      if (api?.query?.broker?.[queryKey]) {
-        try {
-          const result = await api.query.broker[queryKey](...queryParams)
-          // Check if the Option type is Some and unwrap the value
-          if (result) {
-            setData(result.toString())
-          } else {
-            setData(null)
-          }
-        } catch (error) {
-          console.error(`Failed to fetch ${queryKey}:`, error)
-        }
-      }
-    }
-
-    fetchData()
-    const intervalId = setInterval(fetchData, 5000)
-
-    return () => clearInterval(intervalId)
-  }, [api, queryKey, queryParams])
-
-  return data
-}
-
-function useCurrentBlockNumber(api: ApiPromise | undefined) {
-  const [currentBlockNumber, setCurrentBlockNumber] = useState(0)
-
-  useEffect(() => {
-    if (!api) return
-
-    const fetchCurrentBlockNumber = async () => {
-      const currentBlock = await getCurrentBlockNumber(api)
-      setCurrentBlockNumber(currentBlock)
-    }
-
-    const intervalId = setInterval(fetchCurrentBlockNumber, 1000) // Update every second
-
-    return () => clearInterval(intervalId)
-  }, [api])
-
-  return currentBlockNumber
-}
-
-function useBrokerConstants(api: ApiPromise | undefined) {
-  const [brokerConstants, setBrokerConstants] = useState<BrokerConstantsType | null>(null)
-  const [isLoading, setIsLoading] = useState(true)
-
-  useEffect(() => {
-    let isMounted = true
-
-    const fetchConstants = async () => {
-      try {
-        const constants = await getConstants(api)
-        if (isMounted) {
-          setBrokerConstants(constants)
-          setIsLoading(false)
-        }
-      } catch (err) {
-        console.error(err)
-        setIsLoading(false)
-      }
-    }
-
-    fetchConstants()
-
-    return () => {
-      isMounted = false
-    }
-  }, [api])
-
-  return { brokerConstants, isLoading }
-}
 
 function calculateCurrentPrice(
   currentBlockNumber: number,
